scripts/seed/requirements.txt
psycopg2-binary>=2.9.9
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
openpyxl>=3.1.0
